feat: setup Gitea CI/CD workflow and project documentation (#22)
- Replace GitHub Actions workflow with Gitea-compatible CI/CD
  (test, version-bump, build-android, build-linux with release uploads)
- Restructure tests into test/core/ and test/features/ for CI
- Populate CLAUDE.md with code conventions and workflow rules
- Populate docs/TECH_STACK.md with full architecture reference
- Populate test/TEST_COVERAGE.md with test inventory
