+++ .gitignore
@@ -67,3 +67,6 @@ pubspec.lock
 devtools_options.yaml
 android/app/src/main/res/values/themes.xml
 .env
+
+# issue files
+*issue_*.md
