+++ test/features/placeholder_test.dart
@@ -1,0 +1,9 @@
+import 'package:flutter_test/flutter_test.dart';
+
+void main() {
+  test('Features test directory placeholder', () {
+    // Placeholder to ensure test/features/ is recognized by flutter test.
+    // Replace with actual feature tests as they are implemented.
+    expect(true, isTrue);
+  });
+}
