--- test/widget_test.dart
@@ -1,42 +1,0 @@
-// This is a basic Flutter widget test for Document Scanner app.
-//
-// To perform an interaction with a widget in your test, use the WidgetTester
-// utility in the flutter_test package. For example, you can send tap and scroll
-// gestures. You can also use WidgetTester to find child widgets in the widget
-// tree, read text, and verify that the values of widget properties are correct.
-
-import 'package:flutter/material.dart';
-import 'package:flutter_test/flutter_test.dart';
-
-void main() {
-  testWidgets('Document Scanner app smoke test', (WidgetTester tester) async {
-    // Build a simple app widget
-    await tester.pumpWidget(const MaterialApp(home: Scaffold(body: Center(child: Text('Document Scanner')))));
-
-    // Verify that the app shows some basic content
-    expect(find.text('Document Scanner'), findsOneWidget);
-  });
-
-  test('Document naming with underscores test', () {
-    // Test our document naming logic changes
-    final now = DateTime.now();
-    final expectedName = 'Scanned_Document_${now.day}_${now.month}_${now.year}';
-
-    // Verify that the name uses underscores instead of spaces
-    expect(expectedName.contains(' '), false);
-    expect(expectedName.contains('_'), true);
-
-    // Verify specific patterns are replaced
-    expect('Scanned Document 1'.replaceAll(' ', '_'), equals('Scanned_Document_1'));
-    expect('Scanned Document 29/5/2025'.replaceAll(' ', '_').replaceAll('/', '_'), equals('Scanned_Document_29_5_2025'));
-  });
-
-  test('Document scanner service exception handling', () {
-    // Test that our service throws proper exceptions instead of falling back
-    const testException = 'Document scanner failed: Test error. Please ensure camera permissions are granted and try again.';
-
-    // Verify that the exception message format is correct
-    expect(testException.contains('Document scanner failed:'), true);
-    expect(testException.contains('Please ensure camera permissions'), true);
-  });
-}
